--- packages/apptive_grid_form/lib/google_maps_web_service/google_maps_web_service.dart
@@ -1,9 +1,0 @@
-/// A local version with the features required to provide google places search and geocoding to form widgets
-///
-/// This is required as the published google_maps_web_service does not offer null safety in a stable release
-/// Once https://github.com/lejard-h/google_maps_webservice/issues/138 is resolved this should be removed and the published version should be used again
-library google_maps_web_service;
-
-export 'package:apptive_grid_form/google_maps_web_service/core.dart';
-export 'package:apptive_grid_form/google_maps_web_service/geocoding.dart';
-export 'package:apptive_grid_form/google_maps_web_service/places.dart';
