+++ packages/apptive_grid_form/lib/google_maps_webservice/google_maps_webservice.dart
@@ -1,0 +1,9 @@
+/// A local version with the features required to provide google places search and geocoding to form widgets
+///
+/// This is required as the published google_maps_webservice does not offer null safety in a stable release
+/// Once https://github.com/lejard-h/google_maps_webservice/issues/138 is resolved this should be removed and the published version should be used again
+library google_maps_webservice;
+
+export 'package:apptive_grid_form/google_maps_webservice/core.dart';
+export 'package:apptive_grid_form/google_maps_webservice/geocoding.dart';
+export 'package:apptive_grid_form/google_maps_webservice/places.dart';
