+++ packages/apptive_grid_form/test/geolocation_form_widget_test.dart
@@ -2,7 +2,7 @@ import 'dart:async';
 import 'dart:convert';
 
 import 'package:apptive_grid_form/apptive_grid_form.dart';
-import 'package:apptive_grid_form/google_maps_web_service/google_maps_web_service.dart';
+import 'package:apptive_grid_form/google_maps_webservice/google_maps_webservice.dart';
 import 'package:apptive_grid_form/widgets/apptive_grid_form_widgets.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
