+++ packages/apptive_grid_form/test/managers/locationn_manager_test.dart
@@ -1,7 +1,7 @@
 import 'dart:convert';
 
 import 'package:apptive_grid_form/apptive_grid_form.dart';
-import 'package:apptive_grid_form/google_maps_web_service/google_maps_web_service.dart';
+import 'package:apptive_grid_form/google_maps_webservice/google_maps_webservice.dart';
 import 'package:apptive_grid_form/managers/location_manager.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:geolocator/geolocator.dart';
